third_party/WebKit/LayoutTests/http/tests/notifications/service-worker-show-notification-click.html

<!doctype html>
<html>
  <head>
    <title>Notifications: ServiceWorkerRegistration.showNotification().</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that the showNotification() function resolves a promise, and that the
      // notificationclick event gets fired on the Service Worker when we simulate a
      // click on it. This test requires the test runner.
      async_test(function (test) {
          var scope = 'resources/scope/service-worker-show-notification-click',
              workerUrl = 'resources/click-forward-service-worker.js';

          testRunner.grantWebNotificationPermission(location.origin, true);

          var registration = null,
              messagePort = null;
          service_worker_unregister_and_register(test, workerUrl, scope).then(function (swRegistration) {
              registration = swRegistration;
              return wait_for_state(test, registration.installing, 'activated');
          }).then(function () {
              assert_not_equals(registration.active, null, 'The Service Worker needs to be activated.');
              return new Promise(function (resolve) {
                  var messageChannel = new MessageChannel();
                  messageChannel.port1.addEventListener('message', function(event) {
                      if (event.data == 'ready')
                          resolve();
                  });

                  registration.active.postMessage(messageChannel.port2, [messageChannel.port2]);

                  messagePort = messageChannel.port1;
                  messagePort.start();
              });
          }).then(function () {
              assert_inherits(registration, 'showNotification', 'showNotification() must be exposed.');
              return registration.showNotification(scope, {
                  body: 'Hello, world!',
                  icon: '/icon.png'
              });
          }).then(function () {
              messagePort.addEventListener('message', function(event) {
                  assert_equals(event.data, 'The notification has been clicked on.');
                  test.done();
              });

              testRunner.simulateWebNotificationClick(scope);

          }).catch(unreached_rejection(test));

      }, 'Clicking on a notification displayed through showNotification() fires a Service Worker event.');
    </script>
  </body>
</html>
